+++ demo/demo.js
@@ -38,12 +38,14 @@
 document.addEventListener( "DOMContentLoaded", function() {
    const rtm = new RealtimeMonitor();
 
+   loadThemeFromCookie();
+
    const panel1 = rtm.newPanel( {
       title  : "GET Polling",
-      url    : { address  : "https://localhost:8081/status/site1",
-                 method   : "GET",
+      url    : { address : "https://localhost:8081/status/site1",
+                 method  : "GET",
                  pollInterval : 4 },
-      notifications  : false,
+      notifications : false,
       fields : [
                   { prop : "load", label : "Load", suffix : "%", highThresholds : { warn : 70, danger : 90 } },
                   { prop : "rpm", label : "RPM",  lowThresholds : { warn : 500, danger : 300 }, highThresholds : { warn : 2000, danger : 2500 } },
@@ -104,4 +106,25 @@ document.addEventListener( "DOMContentLoaded", function() {
    document.body.appendChild( panel2 );
    document.body.appendChild( panel3 );
    document.body.appendChild( panel4 );
+
+   document.addEventListener( "RTMThemeChange", function(event) {
+      saveThemeToCookie( event.detail.themeName );
+   } );
+
+   function saveThemeToCookie( themeName ) {
+      const expires = new Date();
+		expires.setTime( expires.getTime() + (365 * 24 * 60 * 60 * 1000) );
+
+      // Could use rtm.getTheme() to get the current theme, but that's not as efficient when
+      // you've already received the theme name from a RTMThemeChange CustomEvent.
+      document.cookie = "rtmTheme=" + themeName + "; expires=" + expires.toGMTString();
+   }
+
+   function loadThemeFromCookie() {
+      const theme = decodeURIComponent( document.cookie ).substr( 9 );
+
+      if( ("" + theme).length > 0 ) {
+         rtm.loadTheme( theme );
+      }
+   }
 } );